do not try to update non go files (#33)
Only valid go files can be parsed, if you have a project with c files
next to the go code the update will fail as it tries to parse all
files not just the go ones.

Trying to parse a c file fail with "illegal character U+0023 '#'" for
example and thus fails the upgrade command.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>